backend/requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database - Python 3.13 compatible versions
sqlalchemy>=2.0.35
psycopg2-binary>=2.9.10
alembic>=1.14.0

# Data Validation - Python 3.13 compatible
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.2.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Excel Processing
pandas>=2.2.0
openpyxl>=3.1.2
xlsxwriter>=3.1.9

# Utilities
python-dateutil>=2.8.2
python-dotenv>=1.0.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
